Add some comments to map layers to paper for pytorch code
https://people.csail.mit.edu/khosla/papers/cvpr2016_Khosla.pdf has some notations regarding the various layers. This adds comments to note those layers in the appropriate places.

It is worth noting that the paper calls out 4 convolutional layers CONV-E1 -> CONV-E4. For the face it calls out convolutional layers CONV-F1 -> CONV-F5 in Figure 5, but the caption only calls out CONV-F1 -> CONV-F4. However, the code itself uses ItrackerImageModel on both the face and the eye image layers, which only has 4 convolutional layers. Thus, it seems as though the CONV-F5 in figure 5's image is an error.